refactor Gruff::Base #setup_graph_measurements to reduce complexity (#444)
* refactor Gruff::Base #setup_graph_measurements

Refactored Gruff::Base #setup_graph_measurements to use private methods to calculate each measurement. This moves the intertwined calculations based on hidden aspects of the graph to separate methods to make the decision points easier to follow.

* fix rescued retry for Gruff::Renderer #gradated_background uses wrong parameters

The parameters for this method changed when it moved from `Gruff::Base #render_gradated_background` to `Gruff::Renderer #gradated_background`.  The rescue retry was not updated to retry with the new set of parameters.  This commit fixes that.

**Related Work:**  PR #357 - refactor: Move setup phase of RMagick object into Gruff::Renderer